@@ -1,39 +1,51 @@
 {
 	"name": "vklin",
-	"version": "0.0.3",
 	"private": true,
+	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --check . && eslint .",
-		"format": "prettier --write ."
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint ."
 	},
 	"devDependencies": {
-		"@iconify-json/mdi": "^1.1.67",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/animejs": "^3.1.12",
-		"@types/eslint": "^9.6.0",
-		"@types/three": "^0.169.0",
+		"@eslint/compat": "^1.2.5",
+		"@eslint/js": "^9.18.0",
+		"@iconify-json/mdi": "^1.2.3",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/kit": "^2.16.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@tailwindcss/typography": "^0.5.16",
+		"@types/animejs": "^3.1.13",
+		"@types/color": "^4.2.0",
+		"@types/three": "^0.172.0",
 		"animejs": "^3.2.2",
-		"eslint": "^9.0.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
-		"globals": "^15.0.0",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"sass": "^1.77.8",
-		"svelte": "^5.0.0-next.1",
-		"svelte-check": "^3.6.0",
-		"three": "^0.169.0",
+		"autoprefixer": "^10.4.20",
+		"bits-ui": "^1.0.0-next.78",
+		"clsx": "^2.1.1",
+		"color": "^5.0.0",
+		"eslint": "^9.18.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.14.0",
+		"lucide-svelte": "^0.475.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.3",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
+		"tailwind-merge": "^2.6.0",
+		"tailwind-variants": "^0.3.1",
+		"tailwindcss": "^3.4.17",
+		"tailwindcss-animate": "^1.0.7",
+		"three": "^0.172.0",
 		"typescript": "^5.0.0",
-		"typescript-eslint": "^8.0.0-alpha.20",
-		"unplugin-icons": "^0.19.0",
-		"vite": "^5.0.3"
-	},
-	"type": "module"
+		"typescript-eslint": "^8.20.0",
+		"unplugin-icons": "^22.0.0",
+		"vite": "^6.0.0"
+	}
 }
