@@ -8,6 +8,7 @@
 	interface Props {
 		options: T[];
 		currentOption: number;
+		prevOption?: number;
 		color?: string;
 		disableOther?: boolean;
 		isDisabled?: boolean;
@@ -17,16 +18,15 @@
 	let {
 		options = [],
 		currentOption = $bindable(0),
-		color,
+		prevOption = $bindable(0),
 		disableOther,
 		isDisabled,
 		onOptionChange
 	}: Props = $props();
+
 	let buttonElements: HTMLButtonElement[] = $state([]);
-	let currentOptionWidth = $derived(
-		buttonElements[currentOption] ? buttonElements[currentOption].offsetWidth : 0
-	);
-	let buttonsStartPositions: number[] = $derived(
+	let currentOptionWidth = $derived.by(() => buttonElements[currentOption]?.offsetWidth || 0);
+	let buttonsStartPositions: number[] = $derived.by(() =>
 		makeAccumulativeArray(buttonElements.map((b) => b.offsetWidth))
 	);
 
@@ -41,22 +41,19 @@
 	}
 
 	function selectOption(optionValue: number) {
+		prevOption = currentOption;
 		currentOption = optionValue;
 
 		if (onOptionChange) onOptionChange(optionValue);
 	}
 </script>
 
-<div
-	class="button-select"
-	class:disabled={isDisabled}
-	style="--selector-color: {color ||
-		'buttonface'}; --selector-size: 50px; --selector-width: {currentOptionWidth}px;"
->
+<div class="button-select bg-background bg-opacity-50" class:disabled={isDisabled}>
 	{#each options as option, idx}
 		<button
 			type="button"
-			class:selected={idx === currentOption}
+			class="data-[selected=true]:text-primary-foreground opacity-50 data-[selected=true]:opacity-100 data-[selected=true]:pointer-events-none]"
+			data-selected={idx === currentOption}
 			class:disabled={disableOther && idx !== currentOption}
 			onclick={() => selectOption(idx)}
 			bind:this={buttonElements[idx]}
@@ -65,18 +62,18 @@
 		</button>
 	{/each}
 	<span
-		class="button-select-selector"
-		style="translate: {buttonsStartPositions[currentOption]}px 0;"
+		class="button-select-selector bg-primary"
+		style="translate: {buttonsStartPositions[currentOption]}px 0;
+		width: {currentOptionWidth}px;"
 	></span>
 </div>
 
-<style lang="scss">
+<style>
 	.button-select {
 		position: relative;
 		display: flex;
 		align-items: center;
-		border-radius: 8px;
-		color: white;
+		border-radius: 0.5rem;
 		overflow: hidden;
 
 		&.disabled {
@@ -84,15 +81,13 @@
 			pointer-events: none;
 		}
 
-		&-selector {
+		.button-select-selector {
 			position: absolute;
 			top: 0;
 			left: 0;
 			z-index: 0;
-			border-radius: 8px 8px 0 0;
-			width: var(--selector-width);
+			border-radius: 0.5rem;
 			height: 100%;
-			background-color: var(--selector-color);
 			transition: 0.3s;
 			pointer-events: none;
 		}
@@ -101,34 +96,19 @@
 			position: relative;
 			z-index: 1;
 			border: none;
-			padding: 8px 16px;
+			padding: 0.5rem 1rem;
 			width: 100%;
 			line-height: 1;
-			font-weight: 500;
+			font-weight: 600;
 			text-transform: uppercase;
+			font-size: 0.875rem;
 			background-color: transparent;
-			color: var(--on-surface);
-			opacity: 0.5;
 			transition: 0.3s;
 			cursor: pointer;
 
 			&.disabled {
 				opacity: 0.3;
 				pointer-events: none;
-			}
-
-			&.selected {
-				opacity: 1;
-				color: black;
-				pointer-events: none;
-			}
-
-			&:last-child {
-				border-radius: 0 8px 8px 0;
-			}
-
-			&:first-child {
-				border-radius: 8px 0 0 8px;
 			}
 
 			&:hover {
