@@ -1,13 +1,14 @@
-import type { GLTF } from "three/examples/jsm/loaders/GLTFLoader.js";
-import type { LocationType, Vector2D } from "./interfaces";
-import * as THREE from 'three';
+import { type ClassValue, clsx } from "clsx";
+import { twMerge } from "tailwind-merge";
+import * as THREE from "three";
+import type { Vector2D } from "./interfaces";
 import { gltfLoader } from "./constants";
+import type { GLTF } from "three/examples/jsm/loaders/GLTFLoader.js";
 
-export function remToPixel(rem: number) {
-  return rem * parseFloat(getComputedStyle(document.body).fontSize);
-}
+export const ARROW_OBJECT: THREE.Group<THREE.Object3DEventMap> = new THREE.Group();
+export const POINTS_OBJECT: THREE.Group<THREE.Object3DEventMap> = new THREE.Group();
 
-export function getLocationColor(cellType: LocationType) {
+export function getLocationColor(cellType: string) {
   switch (cellType) {
     case 'drain': {
       return '#ff0000';
@@ -33,8 +34,20 @@ export function getLocationColor(cellType: LocationType) {
   }
 }
 
-export function isSamePosition(a: Vector2D, b: Vector2D) {
-  return a.x === b.x && a.y === b.y
+export function loadAudio(audioFile: string) {
+  const audio = new Audio(audioFile);
+
+  audio.load();
+
+  return new Audio(audioFile);
+}
+
+export async function playSound(sound: HTMLAudioElement, volume: number = 0.1, loop = false) {
+  sound.volume = volume;
+  sound.pause();
+  sound.currentTime = 0;
+  sound.loop = loop;
+  sound.play();
 }
 
 export function randomWithinRange(max: number): number {
@@ -60,20 +73,13 @@ export function getNoun(num: number, words: string[]) {
   const lastDigit = lastTwoDigits % 10;
 
   if (lastTwoDigits >= 11 && lastTwoDigits <= 14) {
-    return words[2];
+    return `${num} ${words[2]}`;
   } else if (lastDigit === 1) {
-    return words[0];
+    return `${num} ${words[0]}`;
   } else if (lastDigit >= 2 && lastDigit <= 4) {
-    return words[1];
+    return `${num} ${words[1]}`;
   } else {
-    return words[2];
-  }
-}
-
-export function subtractVectors(a: Vector2D, b: Vector2D) {
-  return {
-    x: a.x - b.x,
-    y: a.y - b.y
+    return `${num} ${words[2]}`;
   }
 }
 
@@ -90,10 +96,72 @@ export function getGroupCenter(group: THREE.Group) {
   return center;
 }
 
+export async function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+export function isSamePosition(a: Vector2D, b: Vector2D) {
+  return a.x === b.x && a.y === b.y
+}
+
+export function subtractVectors(a: Vector2D, b: Vector2D) {
+  return {
+    x: a.x - b.x,
+    y: a.y - b.y
+  }
+}
+
 export async function loadGLB(glb: string) {
   return new Promise<GLTF>((resolve) => {
     gltfLoader.load(glb, async (gltf) => {
       resolve(gltf);
     });
   });
+}
+
+export function createDiceMesh(sides: number, meshColor = '#000') {
+  const geometry = getDiceGeometry(sides);
+  const color = new THREE.Color(meshColor);
+  const material = new THREE.MeshPhongMaterial({ color })
+  const mesh = new THREE.Mesh(geometry, material);
+
+  return new THREE.Group().add(mesh);
+}
+
+export function getDiceGeometry(sides: number) {
+  switch (sides) {
+    case 4: {
+      return new THREE.TetrahedronGeometry();
+    }
+    case 6: {
+      return new THREE.BoxGeometry();
+    }
+    case 8: {
+      return new THREE.OctahedronGeometry();
+    }
+    case 12: {
+      return new THREE.DodecahedronGeometry();
+    }
+    case 20: {
+      return new THREE.IcosahedronGeometry();
+    }
+  }
+}
+
+export function getLocalTime(timestamp: number) {
+  return new Date(timestamp).toLocaleString('ru-RU', {
+    month: 'long',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit'
+  })
+}
+
+export function remToPixel(rem: number) {
+  return rem * parseFloat(getComputedStyle(document.body).fontSize);
+}
+
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
 }
