@@ -2,36 +2,32 @@
 	import { onMount } from 'svelte';
 	import LastLogs from '../lib/components/LastLogs.svelte';
 	import PlayersBanners from '../lib/components/PlayersBanners.svelte';
-	import Spinner from '../lib/components/Spinner.svelte';
-	import type {
-		FirebaseTreasure,
-		LocationsCollection,
-		LogsCollection,
-		PlayersCollection
-	} from '../lib/interfaces';
-	import Legend from '$lib/components/Legend.svelte';
+	import type { LocationsCollection, LogsCollection, PlayersCollection } from '../lib/interfaces';
 	import type { GLTF } from 'three/examples/jsm/loaders/GLTFLoader.js';
-	import mapController from '$lib/MapController.svelte';
-	import { loadGLB } from '$lib/utils';
+	import { ARROW_OBJECT, loadGLB, POINTS_OBJECT } from '$lib/utils';
 	import arrowGLB from '$lib/assets/3d-models/arrow/direction_arrow.glb';
-	import characterGLB from '$lib/assets/3d-models/character/character.glb';
-	import treasureGLB from '$lib/assets/3d-models/treasure/treasure.glb';
+	import characterGLB from '$lib/assets/3d-models/character/archiedos.glb';
+	import locationsGLB from '$lib/assets/3d-models/locations.glb';
+	import mapGLB from '$lib/assets/3d-models/map/map.glb';
 	import * as SkieletonUtils from 'three/addons/utils/SkeletonUtils.js';
 	import ThreePlayer from '$lib/threeEntities/ThreePlayer';
-	import GameCanvas from '$lib/components/GameCanvas.svelte';
+	import Spinner from '$lib/components/Spinner.svelte';
+	import gameController from '$lib/GameController.svelte';
+	import SpectatorCanvas from '$lib/components/SpectatorCanvas.svelte';
+	import Rules from '$lib/components/Rules.svelte';
+	import GamePlayer from '$lib/gameEntities/GamePlayer.svelte';
+	import gameLogs from '$lib/stores/gameLogs';
+	import itemManager from '$lib/ItemManager';
+	import effectManager from '$lib/effectManager';
+	import EFFECTS from '$lib/data/effectsData';
+	import ITEMS from '$lib/data/itemsData';
 
 	let isLoading = $state(true);
 	let players: PlayersCollection = $state({});
 	let locations: LocationsCollection = $state({});
-	let logs: LogsCollection = $state({});
-	let treasure: FirebaseTreasure = $state({
-		position: { x: 0, y: 0, z: 0 },
-		keysRequired: 0
-	});
 
 	let gltfCharacter: GLTF;
-	let gltfTreasure: GLTF;
-	let gltfArrow: GLTF;
+	let gltfMap: GLTF;
 
 	onMount(async () => {
 		isLoading = true;
@@ -41,37 +37,52 @@
 
 		createLocations(locations);
 		createPlayers(players);
-		createTreasure(treasure);
+		effectManager.addEffects(EFFECTS);
+		itemManager.addItems(ITEMS);
 
 		isLoading = false;
 	});
 
 	async function load3DModels() {
-		const [arrow, character, treasure] = await Promise.all([
+		const [arrow, map, points, character] = await Promise.all([
 			loadGLB(arrowGLB),
-			loadGLB(characterGLB),
-			loadGLB(treasureGLB)
+			loadGLB(mapGLB),
+			loadGLB(locationsGLB),
+			loadGLB(characterGLB)
 		]);
 
-		gltfArrow = arrow;
-		gltfTreasure = treasure;
+		points.scene.children.forEach((child) => {
+			child.children[0].name = 'hexagon';
+		});
+
+		POINTS_OBJECT.add(points.scene);
+		ARROW_OBJECT.add(arrow.scene);
 		gltfCharacter = character;
+		gltfMap = map;
+
+		gameController.addMap(gltfMap.scene);
 	}
 
 	async function fetchGameData() {
-		const data = await fetch('/api/game-data').then((res) => res.json());
+		const res = await fetch('/api/game-data');
+		const data = await res.json();
 
 		players = data.players;
 		locations = data.locations;
-		treasure = data.treasure;
-		logs = data.logs;
+		gameLogs.set(data.logs);
 	}
 
 	function createLocations(locationsValue: LocationsCollection) {
 		for (const location of Object.values(locationsValue)) {
-			const ownerColor = location.ownerId ? players[location.ownerId]?.color : undefined;
+			const ownerColor = location.ownerId
+				? gameController.getPlayerById(location.ownerId)?.color
+				: undefined;
 
-			mapController.addLocation(location, gltfArrow, ownerColor);
+			gameController.addLocation(location, ownerColor);
+		}
+
+		for (const gameLocation of gameController.gameLocations) {
+			gameController.addArrows(gameLocation);
 		}
 	}
 
@@ -82,47 +93,30 @@
 			const threePlayer = new ThreePlayer({
 				...player,
 				object3d: SkieletonUtils.clone(gltfCharacter.scene),
-				animations: gltfCharacter.animations,
-				direction: locations[`${player.position.x}-${player.position.y}`].possibleMoves[0]
+				animations: gltfCharacter.animations
 			});
 
-			mapController.addPlayer(threePlayer);
-		}
-	}
+			let gamePlayer: GamePlayer = new GamePlayer(player, threePlayer);
 
-	function createTreasure(treasureValue: FirebaseTreasure) {
-		mapController.createTreasure(gltfTreasure);
+			gameController.addPlayer(gamePlayer);
+		}
 	}
 </script>
 
-<div class="game-page">
-	{#if isLoading}
-		<div class="loading-page">
+<div class="w-full h-full">
+	{#if isLoading || !players}
+		<div class="w-full h-full flex items-center justify-center">
 			<Spinner --spinner-size="3rem" />
 		</div>
 	{:else}
-		<GameCanvas
-			locationsGroup={mapController.locationsObjects}
-			playersGroup={mapController.playersObjects}
-			treasure={mapController.treasureObject}
+		<SpectatorCanvas
+			locationsGroup={gameController.locationsGroup}
+			playersGroup={gameController.playersGroup}
 		/>
-		<PlayersBanners {players} {logs} />
-		<LastLogs {logs} />
-		<Legend {locations} />
+		<PlayersBanners />
+		<div class="absolute top-[1rem] right-[1rem] z-50 flex gap-2">
+			<Rules />
+			<LastLogs />
+		</div>
 	{/if}
 </div>
-
-<style lang="scss">
-	.game-page {
-		width: 100%;
-		height: 100%;
-	}
-
-	.loading-page {
-		width: 100%;
-		height: 100%;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-	}
-</style>
